Extracted midi controller hook into custom hook
